Add events object, handle irregular disconnects
Socket IDs are now saved in connection with the user object they belong to, as to handle irregular disconnects that occur when the "logout" event is not emitted by the client.

In addition, an "events" object has been created which contains all of the event functions -- this is because I would have otherwise made a seperate function for the logout and disconnect event alone, as they use the same code, and if I had done that then that would have introduced inconsistency with the other methods.

So I decided to simply handle all event functions this way.